7주차 - header 영역 컴포넌트 분리
