+++ src/components/LoginButton.js
@@ -1,0 +1,5 @@
+const LoginButton = () => {
+  return <button className="login_btn">로그인</button>;
+};
+
+export default LoginButton;
